Fix badge layout bug
Moved badge rendering into separate library so that the command line tool and the function can share the code.

Related issue: #6
