+++ Solutions/Endjin.Badger/ColorScheme.cs
@@ -1,0 +1,17 @@
+﻿namespace Endjin.Badger
+{
+    public static class ColorScheme
+    {
+        public static string BrightGreen = "#4c1";
+        public static string Green = "#97CA00";
+        public static string Yellow = "#dfb317";
+        public static string YellowGreen = "#a4a61d";
+        public static string Orange = "#fe7d37";
+        public static string Red = "#e05d44";
+        public static string Blue = "#007ec6";
+        public static string Grey = "#555";
+        public static string Gray = "#555";
+        public static string LightGrey = "#9f9f9f";
+        public static string LightGray = "#9f9f9f";
+    }
+}
