+++ Solutions/Endjin.Badger/Style.cs
@@ -1,0 +1,9 @@
+﻿namespace Endjin.Badger
+{
+    public enum Style
+    {
+        Flat,
+        FlatSquare,
+        Plastic
+    }
+}
